Revision operations in its own client
