+++ apg-patch-service-cmdclient/src/test/groovy/com/apgsga/patch/service/client/IntegrationTest.groovy
@@ -16,6 +16,7 @@ import com.apgsga.microservice.patch.api.Patch
 import com.apgsga.microservice.patch.api.PatchPersistence
 import com.apgsga.microservice.patch.api.ServicesMetaData
 import com.apgsga.microservice.patch.server.MicroPatchServer;
+import com.apgsga.patch.service.client.revision.PatchRevisionCli
 import com.fasterxml.jackson.databind.ObjectMapper
 import groovy.json.JsonBuilder
 import groovy.json.JsonSlurper
@@ -322,176 +323,6 @@ public class IntegrationTest extends Specification {
 		result.returnCode == 0
 	}
 	
-	def "Patch Cli validate retrieve and save revision"() {
-		setup:
-			def client = PatchCli.create()
-			PrintStream oldStream
-			def buffer
-			def revisionAsJson
-			def revisionsFromRRCall
-			def revisionsFromFile
-			def revisionsFile = new File("src/test/resources/Revisions.json")
-		when:
-			oldStream = System.out;
-			buffer = new ByteArrayOutputStream()
-			System.setOut(new PrintStream(buffer))
-			client.process(["-rr", "T,CHEI212"])
-			System.setOut(oldStream)
-			revisionAsJson = getRetrieveRevisionLine(buffer.toString())
-			revisionsFromRRCall = new JsonSlurper().parseText(revisionAsJson)
-		then:
-			revisionsFromRRCall.fromRetrieveRevision.revision.toInteger() == 10000
-			revisionsFromRRCall.fromRetrieveRevision.lastRevision == "SNAPSHOT"
-			!revisionsFile.exists()
-		when:
-			client.process(["-sr", "T,CHEI212,${revisionsFromRRCall.fromRetrieveRevision.revision}"])
-			revisionsFromFile = new JsonSlurper().parseText(revisionsFile.text)
-		then: 
-			revisionsFile.exists()
-			revisionsFromFile.lastRevisions["CHEI212"].toInteger() == 10000
-			revisionsFromFile.lastRevisions["CHEI211"] == null
-			revisionsFromFile.currentRevision["P"].toInteger() == 1
-			revisionsFromFile.currentRevision["T"].toInteger() == 20000
-		when:
-			oldStream = System.out;
-			buffer = new ByteArrayOutputStream()
-			System.setOut(new PrintStream(buffer))
-			client.process(["-rr", "T,CHEI211"])
-			System.setOut(oldStream)
-			revisionAsJson = getRetrieveRevisionLine(buffer.toString())
-			revisionsFromRRCall = new JsonSlurper().parseText(revisionAsJson)
-			revisionsFromFile = new JsonSlurper().parseText(revisionsFile.text)
-		then:
-			revisionsFromRRCall.fromRetrieveRevision.revision.toInteger() == 20000
-			revisionsFromRRCall.fromRetrieveRevision.lastRevision == "SNAPSHOT"
-			revisionsFile.exists()
-			revisionsFromFile.lastRevisions["CHEI212"].toInteger() == 10000
-			revisionsFromFile.lastRevisions["CHEI211"] == null
-			revisionsFromFile.currentRevision["P"].toInteger() == 1
-			revisionsFromFile.currentRevision["T"].toInteger() == 20000
-		when:
-			client.process(["-sr", "T,CHEI211,${revisionsFromRRCall.fromRetrieveRevision.revision}"])
-			revisionsFromFile = new JsonSlurper().parseText(revisionsFile.text)
-		then:
-			revisionsFile.exists()
-			revisionsFromFile.lastRevisions["CHEI212"].toInteger() == 10000
-			revisionsFromFile.lastRevisions["CHEI211"].toInteger() == 20000
-			revisionsFromFile.currentRevision["P"].toInteger() == 1
-			revisionsFromFile.currentRevision["T"].toInteger() == 30000
-		when:
-			oldStream = System.out;
-			buffer = new ByteArrayOutputStream()
-			System.setOut(new PrintStream(buffer))
-			client.process(["-rr", "T,CHEI212"])
-			System.setOut(oldStream)
-			revisionAsJson = getRetrieveRevisionLine(buffer.toString())
-			revisionsFromRRCall = new JsonSlurper().parseText(revisionAsJson)
-		then:
-			revisionsFromRRCall.fromRetrieveRevision.revision.toInteger() == 10001
-			revisionsFromRRCall.fromRetrieveRevision.lastRevision.toInteger() == 10000
-			revisionsFile.exists()
-		when:
-			client.process(["-sr", "T,CHEI212,${revisionsFromRRCall.fromRetrieveRevision.revision}"])
-			revisionsFromFile = new JsonSlurper().parseText(revisionsFile.text)
-		then:
-			revisionsFile.exists()
-			revisionsFromFile.lastRevisions["CHEI212"].toInteger() == 10001
-			revisionsFromFile.lastRevisions["CHEI211"].toInteger() == 20000
-			revisionsFromFile.currentRevision["P"].toInteger() == 1
-			revisionsFromFile.currentRevision["T"].toInteger() == 30000
-		when:
-			oldStream = System.out;
-			buffer = new ByteArrayOutputStream()
-			System.setOut(new PrintStream(buffer))
-			client.process(["-rr", "P,CHPI211"])
-			System.setOut(oldStream)
-			revisionAsJson = getRetrieveRevisionLine(buffer.toString())
-			revisionsFromRRCall = new JsonSlurper().parseText(revisionAsJson)
-		then:
-			revisionsFromRRCall.fromRetrieveRevision.revision.toInteger() == 1
-			revisionsFromRRCall.fromRetrieveRevision.lastRevision == "SNAPSHOT" 
-			revisionsFile.exists()
-		when:
-			client.process(["-sr", "P,CHPI211,${revisionsFromRRCall.fromRetrieveRevision.revision}"])
-			revisionsFromFile = new JsonSlurper().parseText(revisionsFile.text)
-		then:
-			revisionsFile.exists()
-			revisionsFromFile.lastRevisions["CHEI212"].toInteger() == 10001
-			revisionsFromFile.lastRevisions["CHEI211"].toInteger() == 20000
-			revisionsFromFile.lastRevisions["CHPI211"].toInteger() == 1
-			revisionsFromFile.currentRevision["P"].toInteger() == 2
-			revisionsFromFile.currentRevision["T"].toInteger() == 30000
-		when:
-			oldStream = System.out;
-			buffer = new ByteArrayOutputStream()
-			System.setOut(new PrintStream(buffer))
-			client.process(["-rr", "T,CHEI211"])
-			System.setOut(oldStream)
-			revisionAsJson = getRetrieveRevisionLine(buffer.toString())
-			revisionsFromRRCall = new JsonSlurper().parseText(revisionAsJson)
-		then:
-			revisionsFromRRCall.fromRetrieveRevision.revision.toInteger() == 20001
-			revisionsFromRRCall.fromRetrieveRevision.lastRevision.toInteger() == 20000
-			revisionsFile.exists()
-		when:
-			client.process(["-sr", "T,CHEI211,${revisionsFromRRCall.fromRetrieveRevision.revision}"])
-			revisionsFromFile = new JsonSlurper().parseText(revisionsFile.text)
-		then:
-			revisionsFile.exists()
-			revisionsFromFile.lastRevisions["CHEI212"].toInteger() == 10001
-			revisionsFromFile.lastRevisions["CHEI211"].toInteger() == 20001
-			revisionsFromFile.lastRevisions["CHPI211"].toInteger() == 1
-			revisionsFromFile.currentRevision["P"].toInteger() == 2
-			revisionsFromFile.currentRevision["T"].toInteger() == 30000
-		when:
-			oldStream = System.out;
-			buffer = new ByteArrayOutputStream()
-			System.setOut(new PrintStream(buffer))
-			client.process(["-rr", "P,CHPI211"])
-			System.setOut(oldStream)
-			revisionAsJson = getRetrieveRevisionLine(buffer.toString())
-			revisionsFromRRCall = new JsonSlurper().parseText(revisionAsJson)
-		then:
-			revisionsFromRRCall.fromRetrieveRevision.revision.toInteger() == 2
-			revisionsFromRRCall.fromRetrieveRevision.lastRevision.toInteger() == 1
-			revisionsFile.exists()
-		when:
-			client.process(["-sr", "P,CHPI211,${revisionsFromRRCall.fromRetrieveRevision.revision}"])
-			revisionsFromFile = new JsonSlurper().parseText(revisionsFile.text)
-		then:
-			revisionsFile.exists()
-			revisionsFromFile.lastRevisions["CHEI212"].toInteger() == 10001
-			revisionsFromFile.lastRevisions["CHEI211"].toInteger() == 20001
-			revisionsFromFile.lastRevisions["CHPI211"].toInteger() == 2
-			revisionsFromFile.currentRevision["P"].toInteger() == 3
-			revisionsFromFile.currentRevision["T"].toInteger() == 30000
-		when:
-			oldStream = System.out;
-			buffer = new ByteArrayOutputStream()
-			System.setOut(new PrintStream(buffer))
-			client.process(["-rr", "T,CHEI213"])
-			System.setOut(oldStream)
-			revisionAsJson = getRetrieveRevisionLine(buffer.toString())
-			revisionsFromRRCall = new JsonSlurper().parseText(revisionAsJson)
-		then:
-			revisionsFromRRCall.fromRetrieveRevision.revision.toInteger() == 30000
-			revisionsFromRRCall.fromRetrieveRevision.lastRevision == "SNAPSHOT"
-			revisionsFile.exists()
-		when:
-			client.process(["-sr", "T,CHEI213,${revisionsFromRRCall.fromRetrieveRevision.revision}"])
-			revisionsFromFile = new JsonSlurper().parseText(revisionsFile.text)
-		then:
-			revisionsFile.exists()
-			revisionsFromFile.lastRevisions["CHEI212"].toInteger() == 10001
-			revisionsFromFile.lastRevisions["CHEI211"].toInteger() == 20001
-			revisionsFromFile.lastRevisions["CHEI213"].toInteger() == 30000
-			revisionsFromFile.lastRevisions["CHPI211"].toInteger() == 2
-			revisionsFromFile.currentRevision["P"].toInteger() == 3
-			revisionsFromFile.currentRevision["T"].toInteger() == 40000
-		cleanup:
-			revisionsFile.delete()
-	}
-	
 	def "Patch Cli validate onClone mechanism"() {
 		setup:
 			/*
@@ -503,6 +334,7 @@ public class IntegrationTest extends Specification {
 			 * 
 			 */
 			def client = PatchCli.create()
+			def revisionClient = PatchRevisionCli.create()
 			def revisionsFile = new File("src/test/resources/Revisions.json")
 			def currentRevision = [P:5,T:30000]
 			def lastRevision = [CHEI212:10036,CHEI211:4,CHEI213:20025]
@@ -520,7 +352,7 @@ public class IntegrationTest extends Specification {
 		when:
 			// This is actually that the "oc" would do
 			crResult = client.process(["-cr", "CHEI212"])
-			resrResult = client.process(["-resr", "CHEI212"])
+			resrResult = revisionClient.process(["-resr", "CHEI212"])
 			revisionsAfterClone = new JsonSlurper().parseText(revisionsFile.text)
 		then:
 			resrResult.returnCode == 0
@@ -535,9 +367,9 @@ public class IntegrationTest extends Specification {
 			oldStream = System.out;
 			buffer = new ByteArrayOutputStream()
 			System.setOut(new PrintStream(buffer))
-			result = client.process(["-rr", "T,CHEI212"])
+			result = revisionClient.process(["-rr", "T,CHEI212"])
 			System.setOut(oldStream)
-			revisionAsJson = getRetrieveRevisionLine(buffer.toString())
+			revisionAsJson = TestUtil.getRevisionLine(buffer.toString())
 			revisionsFromRRCall = new JsonSlurper().parseText(revisionAsJson)
 		then:
 			result.returnCode == 0
@@ -545,7 +377,7 @@ public class IntegrationTest extends Specification {
 			revisionsFromRRCall.fromRetrieveRevision.revision.toInteger() == 10000
 			revisionsFromRRCall.fromRetrieveRevision.lastRevision == "CLONED"
 		when:
-			result = client.process(["-sr", "T,CHEI212,${revisionsFromRRCall.fromRetrieveRevision.revision}"])
+			result = revisionClient.process(["-sr", "T,CHEI212,${revisionsFromRRCall.fromRetrieveRevision.revision}"])
 			revisionsFromFile = new JsonSlurper().parseText(revisionsFile.text)
 		then:
 			revisionsFile.exists()
@@ -558,9 +390,9 @@ public class IntegrationTest extends Specification {
 			oldStream = System.out;
 			buffer = new ByteArrayOutputStream()
 			System.setOut(new PrintStream(buffer))
-			result = client.process(["-rr", "T,CHEI212"])
+			result = revisionClient.process(["-rr", "T,CHEI212"])
 			System.setOut(oldStream)
-			revisionAsJson = getRetrieveRevisionLine(buffer.toString())
+			revisionAsJson = TestUtil.getRevisionLine(buffer.toString())
 			revisionsFromRRCall = new JsonSlurper().parseText(revisionAsJson)
 		then:
 			result.returnCode == 0
@@ -574,7 +406,7 @@ public class IntegrationTest extends Specification {
 			revisions = [lastRevisions:lastRevision, currentRevision:currentRevision]
 			revisionsFile.write(new JsonBuilder(revisions).toPrettyString())
 			crResult = client.process(["-cr", "CHQI567"])
-			resrResult = client.process(["-resr", "CHQI567"])
+			resrResult = revisionClient.process(["-resr", "CHQI567"])
 			def revisionsAfterUnvalidClone = new JsonSlurper().parseText(revisionsFile.text)
 		then:
 			result.returnCode == 0
@@ -586,72 +418,5 @@ public class IntegrationTest extends Specification {
 			revisionsAfterUnvalidClone.lastRevisions["CHEI212"].toInteger() == 10036
 		cleanup:
 			revisionsFile.delete()
-	}
-	
-	def "Patch Cli validate retrieve last prod revision"() {
-		setup:
-			/*
-			 * For our tests, within src/test/resources/TargetSystemMappings.json, CHEI211 is configured as the
-			 * production target.
-			 *
-			 */
-			def client = PatchCli.create()
-			def revisionsFile = new File("src/test/resources/Revisions.json")
-			def currentRevision = [P:5,T:30000]
-			def lastRevision = [CHEI212:10036,CHEI211:4,CHEI213:20025]
-			def revisions = [lastRevisions:lastRevision, currentRevision:currentRevision]
-			revisionsFile.write(new JsonBuilder(revisions).toPrettyString())
-			def oldStream
-			def buffer
-			def revisionAsJson
-			def revisionsFromRRCall
-		when:
-			oldStream = System.out;
-			buffer = new ByteArrayOutputStream()
-			System.setOut(new PrintStream(buffer))
-			client.process(["-pr"])
-			System.setOut(oldStream)
-			revisionAsJson = getLastProdRevisionLine(buffer.toString())
-			revisionsFromRRCall = new JsonSlurper().parseText(revisionAsJson)
-		then:
-			revisionsFromRRCall.lastProdRevision.toInteger() == 4
-		cleanup:
-			revisionsFile.delete()
-	}
-	
-	def "Patch Cli delete all T revision with dryRun"() {
-		setup:
-			def client = PatchCli.create()
-		when:
-			client.process(["-rtr", "1"]) // 1 -> dryRun
-		then:
-			// Simply nothing should happen.
-			notThrown(RuntimeException)
-	}
-	
-	def getLastProdRevisionLine(String lines) {
-		// Looking for the line which is for us interesting -> should contain "fromRetrieveRevision"
-		def searchedLine = null
-		lines.eachLine{ line ->
-			if (line != null) {
-				if(line.contains("lastProdRevision")) {
-					searchedLine = line
-				}
-			}
-		}
-		return searchedLine
-	}
-	
-	def getRetrieveRevisionLine(String lines) {
-		// Looking for the line which is for us interesting -> should contain "fromRetrieveRevision"
-		def searchedLine = null
-		lines.eachLine{ line ->
-			if (line != null) {
-				if(line.contains("fromRetrieveRevision")) {
-					searchedLine = line
-				}
-			}
-		}
-		return searchedLine
 	}
 }
